requirements.txt
asgiref==3.5.2
Django==4.1.4
factory-boy==3.2.1
Faker==15.3.4
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
